Basic concept
Basic concept is implemented, but practical implementation of this project has not yet been finished. There are MANY things to tidy up, but I made this once when I was bored, got the basic gist working and then moved onto a proper project. Maybe I'll come back and work on it some more when I'm bored again.